Adding test for requesting from API, also some dagger hook up.
