+++ app/src/main/java/com/adammcneilly/wom/WiseOldManApp.kt
@@ -1,0 +1,7 @@
+package com.adammcneilly.wom
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class WiseOldManApp : Application()
